runtime(Move R.1): reconcile csv unhashable-key drift across in-tree↔satellite copies
The default native build compiles the molt-runtime-serial SATELLITE copy of csv;
the reduced tiers (micro/edge/WASM) compile the IN-TREE copy. The two had drifted
into a real shipped-behavior bug differing by build tier:

  * in-tree builtins/csv.rs::molt_csv_dialect_lookup_name routes a non-string
    dialect name through the GENERAL ensure_hashable(_py, name_bits,
    HashContext::DictKey) check — catches every unhashable type and emits the
    CPython-version-gated "cannot use 'X' as a dict key (unhashable type: 'X')"
    message (bare on 3.12/3.13). Correct.
  * the satellite hardcoded `if tid == LIST|DICT|SET` and a non-version-gated
    message — so csv.get_dialect(bytearray(b"x")) (and every other unhashable
    type) returned csv.Error("unknown dialect") on the DEFAULT native build
    instead of the CPython TypeError.

Reconcile by making the satellite embody the in-tree behavior, using the same
single source of truth (the in-tree ensure_hashable), exposed across the FFI
bridge:
  * add `ensure_hashable: fn(u64, i32) -> i32` + `HashContextCode` (stable repr
    discriminant) to molt_runtime_core::RuntimeVtable + prelude;
  * wire bridge_ensure_hashable in molt-runtime/src/serial_bridge.rs (maps the
    discriminant to HashContext, delegates to object::ops_hash::ensure_hashable);
  * add bridge::ensure_hashable to molt-runtime-serial and replace the hardcoded
    csv check with ensure_hashable(_py, name_bits, HashContextCode::DictKey).

Differential regression tests/differential/stdlib/csv_get_dialect_unhashable.py
pins list/dict/set/bytearray/unknown-str — byte-identical vs CPython 3.14 on BOTH
the default (satellite) and `MOLT_DIFF_STDLIB_PROFILE=micro` (in-tree) tiers.

Also clean a pre-existing redundant `as i32` cast (clippy unnecessary_cast) on
`parse_digits_n(...)?` in BOTH datetime copies symmetrically, so the parity
guard's normalized residual is unchanged and the serial satellite passes
`cargo clippy -- -D warnings`.

+++ runtime/molt-runtime-serial/src/datetime.rs
@@ -573,7 +573,7 @@ fn parse_isotime_str(s: &str) -> Result<(i32, i32, i32, i32, i64), String> {
     if b.len() < 2 {
         return Err(format!("Invalid time string: {s:?}"));
     }
-    let h = parse_digits_n(b, 0, 2)? as i32;
+    let h = parse_digits_n(b, 0, 2)?;
     let mut pos = 2usize;
     let (mi, sec, us) = if pos < b.len() && (b[pos] == b':' || b[pos].is_ascii_digit()) {
         let has_colon = b[pos] == b':';
@@ -583,7 +583,7 @@ fn parse_isotime_str(s: &str) -> Result<(i32, i32, i32, i32, i64), String> {
         if pos + 2 > b.len() {
             return Err(format!("Invalid time string: {s:?}"));
         }
-        let mi = parse_digits_n(b, pos, 2)? as i32;
+        let mi = parse_digits_n(b, pos, 2)?;
         pos += 2;
         let (sec, us) =
             if pos < b.len() && (b[pos] == b':' || (!has_colon && b[pos].is_ascii_digit())) {
@@ -593,7 +593,7 @@ fn parse_isotime_str(s: &str) -> Result<(i32, i32, i32, i32, i64), String> {
                 if pos + 2 > b.len() {
                     return Err(format!("Invalid time string: {s:?}"));
                 }
-                let sec = parse_digits_n(b, pos, 2)? as i32;
+                let sec = parse_digits_n(b, pos, 2)?;
                 pos += 2;
                 let us = if pos < b.len() && b[pos] == b'.' {
                     pos += 1;
